+++ numba/hsa/dispatch.py
